Implemented PlayerPrefs
Using playerprefs, now the player needs to finish first level to unlock the second and so on. Also, now the game keeps track of the collectibles collected at the end of each level and transfers only the ones when player reaches victory. And, I have implemented collectible dropper that drops the collectibles on random chance.